moved generated code outside src to improve code coverage
