+++ .github/workflows/go.yml
@@ -27,7 +27,7 @@ jobs:
       #   run: go test -v ./...
 
       - name: Run Coverage
-        run: go test ./src/... -coverpkg=./src/... -race -covermode=atomic -coverprofile=coverage.out
+        run: go test ./test -coverpkg=./src/... -race -covermode=atomic -coverprofile=coverage.out
 
       - name: Codecov
         uses: codecov/codecov-action@v3
