+++ test/info_test.go
